# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GamingCorner API
  version: "1.0.0"
  description: API para autenticación, catálogo de juegos, ratings, comentarios, géneros e imágenes.
servers:
  - url: http://localhost:3000/api
tags:
  - name: Auth
  - name: Public
  - name: User
  - name: Admin
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
    UserSummary:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role:
          type: string
          enum: [user, admin]
    AuthRegisterReq:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: nico }
        password: { type: string, example: "123456" }
    AuthLoginReq:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: nico }
        password: { type: string, example: "123456" }
    AuthLoginRes:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/UserSummary"
    AuthResetReq:
      type: object
      required: [username, newPassword]
      properties:
        username: { type: string }
        newPassword: { type: string }
    Genre:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, example: JRPG }
    GenreCreateReq:
      type: object
      required: [name]
      properties:
        name: { type: string }
    Game:
      type: object
      properties:
        id: { type: integer }
        game_code: { type: string, example: SWI-TOTK-2023 }
        title: { type: string }
        slug: { type: string }
        platform:
          type: string
          enum: [PC, PS, XBOX, SWI, MBL]
        developer: { type: string, example: "Aggro Crab, Landfall" }
        publisher: { type: string, example: "Aggro Crab, Landfall" }
        release_date: { type: string, format: date }
        trailer_url:
          { type: string, example: "https://www.youtube.com/embed/..." }
        hero_image:
          { type: string, example: "/uploads/SWI-TOTK-2023/portada.jpg" }
        description: { type: string }
        steam_app_id: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        avg_rating:
          type: number
          format: float
          nullable: true
        ratings:
          type: integer
          nullable: true
    GameCreateOrUpdateReq:
      type: object
      required: [title, platform, release_date]
      properties:
        title: { type: string }
        platform:
          type: string
          enum: [PC, PS, XBOX, SWI, MBL]
        release_date: { type: string, format: date, nullable: true }
        developer: { type: string, description: "Separados por coma" }
        publisher: { type: string, description: "Separados por coma" }
        trailer_url: { type: string, nullable: true }
        description: { type: string, nullable: true }
        steam_app_id: { type: integer, nullable: true }
        genre_ids:
          type: array
          items: { type: integer }
    Comment:
      type: object
      properties:
        id: { type: integer }
        content: { type: string }
        created_at: { type: string, format: date-time }
        username: { type: string }
    CommentCreateReq:
      type: object
      required: [content]
      properties:
        content: { type: string }
    RatingReq:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
    ImageURLs:
      type: array
      items:
        type: string
        example: "/uploads/PC-AEDE-2019/portada.jpg"
    PatchNotes:
      oneOf:
        - type: object
          properties:
            title: { type: string }
            contents: { type: string }
            date:
              {
                type: integer,
                description: "Unix timestamp (epoch seconds) de Steam",
              }
        - type: object
          properties:
            note:
              type: string
              nullable: true
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Crear usuario
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthRegisterReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
        "400":
          {
            description: Error,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "409":
          {
            description: Duplicado,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthLoginReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthLoginRes" }
        "401":
          {
            description: Credenciales inválidas,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /auth/reset:
    post:
      tags: [Auth]
      summary: Restablecer contraseña (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthResetReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
        "404":
          {
            description: No encontrado,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /genres:
    get:
      tags: [Public]
      summary: Listar géneros (público)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Genre" }

  /games:
    get:
      tags: [Public]
      summary: Buscar/Listar juegos
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: genre
          schema: { type: string, description: "Nombre del género exacto" }
        - in: query
          name: platform
          schema: { type: string, enum: [PC, PS, XBOX, SWI, MBL] }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 12, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Game" }

  /games/top:
    get:
      tags: [Public]
      summary: Top juegos por rating
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Game" }

  /games/search-codes:
    get:
      tags: [Public]
      summary: Autocomplete por código o título
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code: { type: string }
                    title: { type: string }

  /games/{idOrCode}:
    get:
      tags: [Public]
      summary: Obtener detalle de juego
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/Game" } },
            }
        "404":
          {
            description: No existe,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /games/{idOrCode}/images:
    get:
      tags: [Public]
      summary: Listar imágenes del juego
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImageURLs" }

  /games/{idOrCode}/comments:
    get:
      tags: [Public]
      summary: Listar comentarios (público)
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Comment" }

    post:
      tags: [User]
      summary: Crear comentario (requiere login)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommentCreateReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
        "400":
          {
            description: Error,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "401":
          {
            description: No autorizado,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "404":
          {
            description: Juego no existe,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /games/{idOrCode}/rating:
    put:
      tags: [User]
      summary: Crear/Actualizar rating (1..5) (requiere login)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RatingReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
        "400":
          {
            description: Error,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "401":
          {
            description: No autorizado,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }
        "404":
          {
            description: Juego no existe,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /games/{idOrCode}/patchnotes:
    get:
      tags: [Public]
      summary: Última nota de Steam (si aplica)
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PatchNotes" }

  /admin/genres:
    get:
      tags: [Admin]
      summary: Listar géneros (admin)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Genre" }
    post:
      tags: [Admin]
      summary: Crear género
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenreCreateReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }

  /admin/genres/{id}:
    delete:
      tags: [Admin]
      summary: Eliminar género
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }

  /admin/games:
    post:
      tags: [Admin]
      summary: Crear juego
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameCreateOrUpdateReq" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  game_code: { type: string }

  /admin/games/{idOrCode}:
    put:
      tags: [Admin]
      summary: Actualizar juego
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameCreateOrUpdateReq" }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
    delete:
      tags: [Admin]
      summary: Eliminar juego
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }
        "404":
          {
            description: No existe,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Error" } },
              },
          }

  /admin/games/{idOrCode}/images:
    post:
      tags: [Admin]
      summary: Subir imágenes (multi)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  description: "Hasta 12 archivos"
                  items:
                    type: string
                    format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  images:
                    $ref: "#/components/schemas/ImageURLs"

  /admin/games/{idOrCode}/images/{file}:
    delete:
      tags: [Admin]
      summary: Eliminar imagen de juego
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: idOrCode
          required: true
          schema: { type: string }
        - in: path
          name: file
          required: true
          schema: { type: string }
          description: "Nombre de archivo (exacto)"
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Ok" } },
              },
          }